:articulated_lorry: ‍Migrate lodash to ^4.0.0
This commit is to migrate the `lodash` dependency from the
(deprecated) `~2.4.1` to the current `^4.0.0`.

As lodash v2.4.x has been deprecated for ages, installing or updating
`awm` would make `npm` print the following warning:

```
npm WARN deprecated lodash-node@2.4.1: This package is discontinued. Use lodash@^4.0.0.
```

For first-time users, this makes for a poor out-of-the-box
experience; for developers, I deem this an ongoing distraction.

To check for incompatibilities, I ran the `lodash-migrate` package
prior to bumping the version number to ^4.0.0. This revealed the
following compatibility issues to v2.4.x, which are all fixed in this
commit:

- lodash v3.0.0 has introduced the `_.trim` method, allowing for
  making `awm`’s array splitting code a bit simpler. (Not an actual
  issue.)

- lodash v3.0.0 has renamed the `_.contains` method to `_.includes`;
  the remaining `_.contains` alias was subsequently removed in
  v4.0.0.

- lodash v4.0.0 has removed the `_.pluck` method because it doubles
  the newly introduced iteratee short-cut semantics, e. g.
  `_.map(packages, 'bundle')`.

This commit also bumps the `plist` dependency to `^2.0.0` because its
older version had an indirect dependency on an outdated lodash
release.

Considering the stable API track record of either module within
major releases, I opted for the `^` modifier in both cases.

See also:

- [Package lodash-migrate on npmjs.com](https://www.npmjs.com/package/lodash-migrate)
- [lodash v3.0.0 » Notable Changes](https://github.com/lodash/lodash/wiki/Changelog/23cf804c63cb1dafb3f47bd99cc7c3d96d42d102#notable-changes-1)
- [lodash v4.0.0 » Compatibility Warnings](https://github.com/lodash/lodash/wiki/Changelog/23cf804c63cb1dafb3f47bd99cc7c3d96d42d102#compatibility-warnings)