:recycle: Bump `request`, lose `node-uuid` dependency
`npm install` keeps telling me:

> `npm WARN deprecated node-uuid@1.4.7: use uuid module instead`

With `awm` depending on `node-uuid` only indirectly, and only through
the `request` package, it is probably sufficient to bump the latter
to `^2.79.0` in order to get rid of `node-uuid`.

Apart from that, staying on top of any networking code [should be
considered good practice
anyway](https://github.com/request/request/issues?q=is%3Aissue%20security);
I figure this justifies a `^` modifier so `awm` users continue
to receive the latest 2.x updates of the `request` package from now
on.